! Make sure a file to be deleted during the install of a package is checked for existence first. (Subs-Package.php) [Bug 2739]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@7696 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ Sources/Subs-Package.php
@@ -1362,9 +1362,16 @@ function parsePackageInfo(&$packageXML, $testing_only = true, $method = 'install
 			deltree($action['filename']);
 		elseif ($action['type'] == 'remove-file')
 		{
-			package_chmod($action['filename']);
+			// Make sure the file exists before deleting it.
+			if (file_exists($action['filename']))
+			{
+				package_chmod($action['filename']);
+				$failure |= !unlink($action['filename']);
+			}
 
-			$failure |= !unlink($action['filename']);
+			// The file that was supposed to be deleted couldn't be found.
+			else
+				$failure = true;
 		}
 	}
